requirements.txt
Django~=5.0.7
djoser~=2.2.3
djangorestframework~=3.15.2